update to make grading scale 0-100 and optional

+++ moonshot/src/api/api_recipe.py
@@ -13,7 +13,7 @@ def api_create_recipe(
     prompt_templates: list[str],
     metrics: list[str],
     attack_modules: list[str],
-    grading_scale: dict[str, list[float]],
+    grading_scale: dict[str, list[int]],
 ) -> None:
     """
     Creates a new recipe and stores it in a JSON file.
@@ -33,7 +33,7 @@ def api_create_recipe(
         prompt_templates (list[str]): A list of prompt templates for generating prompts.
         metrics (list[str]): A list of metrics to evaluate the recipe's performance.
         attack_modules (list[str]): A list of attack modules to test the recipe's robustness.
-        grading_scale (dict[list[float]]): A dictionary defining the grading scale with keys as grade labels and values
+        grading_scale (dict[list[int]]): A dictionary defining the grading scale with keys as grade labels and values
         as score ranges.
 
     Returns:
